+++ examples/preload/idleDetector/filterMidiEvent.js
@@ -1,0 +1,11 @@
+function messageEventFilter(messageEvent) {
+  const command = messageEvent.data[0];
+  if (command === 128) {
+    console.log("Ignoring noteOff MIDI message event.", messageEvent);
+    return true;
+  } else {
+    return false;
+  }
+}
+
+kioskBrowser.idleDetector.midi.messageEventFilters.add(messageEventFilter);